feat!: add KomiHasher that Hasher, which is no longer implemented by StreamedKomiHasher
The `Hasher` trait explicitly states that different call to write() should
not be treated as if the bytes were concatenated, doing so is even
dangerous as it made it easy to construct a collision.

So this commit introduce `KomiHasher` that implments `Hasher` in a safe
way. For that reason, StreamedKomiHasher no longer implements Hasher to avoid
being confused.